refactor: Consolidate Main page
