+++ frontend/src/features/main/components/index.ts
@@ -1,1 +1,4 @@
-export { default as MainContent } from './MainContent';
+export { default as MainBrand } from './MainBrand';
+export { default as MainCategoryList } from './MainCategoryList';
+export { default as MainHeroSection } from './MainHeroSection';
+export { default as MainProductList } from './MainProductList';
