+++ frontend/src/features/main/types/MainBrandProps.ts
@@ -1,0 +1,9 @@
+export interface MainBrand {
+    id: number;
+    name: string;
+    logo: string;
+}
+
+export interface MainBrandProps {
+    brands: MainBrand[];
+}
