+++ frontend/src/features/main/types/MainCategoryListProps.ts
@@ -1,0 +1,5 @@
+import type { MainCategory } from "./MainCategoryProps";
+
+export interface MainCategoryListProps {
+    categories: MainCategory[];
+}
