+++ frontend/src/features/main/types/MainCategoryProps.ts
@@ -1,0 +1,10 @@
+export interface MainCategory {
+    name: string;
+    icon: string;
+    count: string;
+}
+
+export interface MainCategoryProps {
+    category: MainCategory;
+    onClick: () => void;
+}
