+++ frontend/src/features/main/types/MainHeroSectionProps.ts
@@ -1,0 +1,3 @@
+export interface MainHeroSectionProps {
+    image: string;
+}
