+++ frontend/src/features/main/types/MainProductListProps.ts
@@ -1,0 +1,5 @@
+import type { MainProduct } from "./MainProductProps";
+
+export interface MainProductListProps {
+    products: MainProduct[];
+}
