+++ frontend/src/features/main/types/MainProductProps.ts
@@ -1,0 +1,12 @@
+export interface MainProduct {
+    id: number;
+    image: string;
+    name: string;
+    price: string;
+    bidCount: number;
+    timeLeft: string;
+}
+
+export interface MainProductProps {
+    product: MainProduct
+}
