--- frontend/src/features/product/components/brand/index.ts
@@ -1,3 +1,0 @@
-export { default as BrandList } from './BrandList';
-export { default as BrandModal } from './BrandModal';
-export { default as BrandPopular } from './BrandPopular';
