--- frontend/src/features/product/types/brand/Brand.ts
@@ -1,21 +1,0 @@
-export interface Brand {
-    id: number;
-    name: string;
-    logo: string;
-}
-
-export interface BrandProduct {
-    id: number;
-    name: string;
-    image: string;
-    currentBid: string;
-    timeLeft: string;
-    bidders: number;
-}
-
-export interface BrandWithProducts {
-    brand: string;
-    description: string;
-    logo: string;
-    products: BrandProduct[];
-}
