--- frontend/src/features/product/types/brand/BrandItemProps.ts
@@ -1,6 +1,0 @@
-import type { BrandProduct } from "./Brand";
-
-export interface BrandItemProps {
-    product: BrandProduct;
-    onBidClick: (product: BrandProduct) => void;
-}
