--- frontend/src/features/product/types/brand/BrandListProps.ts
@@ -1,6 +1,0 @@
-import type { BrandWithProducts, BrandProduct } from "./Brand";
-
-export interface BrandListProps {
-    brandData: BrandWithProducts;
-    onBidClick: (product: BrandProduct) => void;
-}
