--- frontend/src/features/product/types/brand/BrandModalProps.ts
@@ -1,8 +1,0 @@
-import type { BrandProduct } from "./Brand";
-
-export interface BrandModalProps {
-    isOpen: boolean;
-    onClose: () => void;
-    product: BrandProduct | null;
-    onSubmit: (productInd: number, bidAmount: number) => void;
-}
