--- frontend/src/features/product/types/brand/BrandPopular.ts
@@ -1,5 +1,0 @@
-import type { Brand } from "./Brand";
-
-export interface BrandPopularProps {
-    brands: Brand[];
-}
